returns both post and friend request in correct format
